refactor: split off BlockLoader support into BlockArcWarcRecordLoader, plain ArcWarcRecordLoader only includes parse_record_stream(), no load() use BlockArcWarcRecordLoader() only when needed for replay
